EpiModel_agg_bd Example output from the EpiModel package for a individual model with birth and death rates.

This data is made from the SIR DCM/ICM example in the EpiModel vignette found here. We run the discrete compartmental model (DCM) for 300 steps and set the birth and death rates (a.rate, ds.rate, di.rate, dr.rate) to .02. This is object is of class icm. The output details can be found with ?EpiModel::icm.

EpiModel_agg_bd

Format

a dcm object from the EpiModel package. This contains the following elements:

param

the epidemic parameters passed into the model through param, with additional parameters added as necessary.

control

the control settings passed into the model through control, with additional controls added as necessary.

epi

a list of data frames, one for each epidemiological output from the model. Outputs for base models always include the size of each compartment, as well as flows in, out of, and between compartments.

Examples

## Look at structure
str(EpiModel_agg_bd)

#> List of 3
#>  $ param  :List of 9
#>   ..$ inf.prob: num 0.2
#>   ..$ act.rate: num 0.8
#>   ..$ rec.rate: num 0.02
#>   ..$ a.rate  : num 0.02
#>   ..$ ds.rate : num 0.02
#>   ..$ di.rate : num 0.02
#>   ..$ dr.rate : num 0.02
#>   ..$ vital   : logi TRUE
#>   ..$ groups  : num 1
#>   ..- attr(*, "class")= chr [1:2] "param.icm" "list"
#>  $ control:List of 11
#>   ..$ type       : chr "SIR"
#>   ..$ nsteps     : num 300
#>   ..$ nsims      : num 1
#>   ..$ rec.rand   : logi TRUE
#>   ..$ a.rand     : logi TRUE
#>   ..$ d.rand     : logi TRUE
#>   ..$ verbose    : logi FALSE
#>   ..$ verbose.int: num 0
#>   ..$ skip.check : logi FALSE
#>   ..$ bi.mods    : chr [1:6] "initialize.FUN" "infection.FUN" "recovery.FUN" "departures.FUN" ...
#>   ..$ user.mods  : chr(0) 
#>   ..- attr(*, "class")= chr [1:2] "control.icm" "list"
#>  $ epi    :List of 10
#>   ..$ s.num  :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: int [1:300] 900 890 867 853 840 831 804 789 776 752 ...
#>   ..$ i.num  :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: int [1:300] 100 106 120 135 147 154 173 194 209 223 ...
#>   ..$ num    :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: int [1:300] 1000 996 992 994 997 1000 995 1008 1012 1005 ...
#>   ..$ r.num  :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: int [1:300] 0 0 5 6 10 15 18 25 27 30 ...
#>   ..$ si.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: num [1:300] 0 9 20 20 18 16 25 29 21 27 ...
#>   ..$ ir.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: num [1:300] 0 0 5 1 4 5 4 7 2 5 ...
#>   ..$ ds.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: num [1:300] 0 19 20 12 14 13 20 9 12 15 ...
#>   ..$ di.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: num [1:300] 0 3 1 4 2 4 2 1 4 8 ...
#>   ..$ dr.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: num [1:300] 0 0 0 0 0 0 1 0 0 2 ...
#>   ..$ a.flow :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ sim1: num [1:300] 0 18 17 18 19 20 18 23 20 18 ...
#>  - attr(*, "class")= chr "icm"

## Look at SIR values
head(do.call('cbind', EpiModel_agg_bd$epi[c("s.num", "i.num", "r.num")]))

#>   sim1 sim1 sim1
#> 1  900  100    0
#> 2  890  106    0
#> 3  867  120    5
#> 4  853  135    6
#> 5  840  147   10
#> 6  831  154   15