EpiModel_det Example output from the EpiModel package for a deterministic model

This data is made from the SIR DCM/ICM example in the EpiModel vignette found here. We run the discrete compartmental model (DCM) for 300 steps and set the birth and death rates (a.rate, ds.rate, di.rate, dr.rate) to zero. This is object is of class dcm. The output details can be found with ?EpiModel::dcm.

EpiModel_det

Format

a dcm object from the EpiModel package. This contains the following elements:

param

the epidemic parameters passed into the model through param, with additional parameters added as necessary.

control

the control settings passed into the model through control, with additional controls added as necessary.

epi

a list of data frames, one for each epidemiological output from the model. Outputs for base models always include the size of each compartment, as well as flows in, out of, and between compartments.

Examples

## Look at structure
str(EpiModel_det)

#> List of 3
#>  $ param  :List of 9
#>   ..$ inf.prob: num 0.2
#>   ..$ act.rate: num 0.8
#>   ..$ rec.rate: num 0.02
#>   ..$ a.rate  : num 0
#>   ..$ ds.rate : num 0
#>   ..$ di.rate : num 0
#>   ..$ dr.rate : num 0
#>   ..$ vital   : logi TRUE
#>   ..$ groups  : num 1
#>   ..- attr(*, "class")= chr [1:2] "param.dcm" "list"
#>  $ control:List of 11
#>   ..$ type      : chr "SIR"
#>   ..$ nsteps    : num 300
#>   ..$ dt        : num 1
#>   ..$ odemethod : chr "rk4"
#>   ..$ dede      : logi FALSE
#>   ..$ new.mod   : NULL
#>   ..$ sens.param: logi TRUE
#>   ..$ print.mod : logi FALSE
#>   ..$ verbose   : logi FALSE
#>   ..$ nruns     : int 1
#>   ..$ timesteps : num [1:300] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..- attr(*, "class")= chr [1:2] "control.dcm" "list"
#>  $ epi    :List of 10
#>   ..$ s.num  :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 900 885 868 849 829 ...
#>   ..$ i.num  :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 100 113 128 143 161 ...
#>   ..$ r.num  :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 0 2.13 4.53 7.24 10.28 ...
#>   ..$ si.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 15.2 16.9 18.6 20.4 22.3 ...
#>   ..$ ir.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 2.13 2.4 2.71 3.04 3.4 ...
#>   ..$ a.flow :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ ds.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ di.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ dr.flow:'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ num    :'data.frame':	300 obs. of  1 variable:
#>   .. ..$ run1: num [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>  - attr(*, "class")= chr "dcm"

## Look at SIR values
head(do.call('cbind', EpiModel_det$epi[c("s.num", "i.num", "r.num")]))

#>       run1     run1      run1
#> 1 900.0000 100.0000  0.000000
#> 2 884.8051 113.0665  2.128420
#> 3 867.9550 127.5131  4.531860
#> 4 849.3629 143.3986  7.238538
#> 5 828.9614 160.7610 10.277653
#> 6 806.7095 179.6116 13.678904