EpiModel_icm Example output from the EpiModel package for a stochastic ICM

This data is made from the SIR DCM/ICM example in the EpiModel vignette found here. We run the individual contact model (ICM) for 300 steps and set the birth and death rates (a.rate, ds.rate, di.rate, dr.rate) to zero. This is object is of class icm. We run the simulation 10 times. The output details can be found with ?EpiModel::icm.

EpiModel_icm

Format

a icm object from the EpiModel package. This contains the following elements:

param

the epidemic parameters passed into the model through param, with additional parameters added as necessary.

control

the control settings passed into the model through control, with additional controls added as necessary.

epi

a list of data frames, one for each epidemiological output from the model. Outputs for base models always include the size of each compartment, as well as flows in, out of, and between compartments.

Examples

## Look at structure
str(EpiModel_icm)

#> List of 3
#>  $ param  :List of 9
#>   ..$ inf.prob: num 0.2
#>   ..$ act.rate: num 0.8
#>   ..$ rec.rate: num 0.02
#>   ..$ a.rate  : num 0
#>   ..$ ds.rate : num 0
#>   ..$ di.rate : num 0
#>   ..$ dr.rate : num 0
#>   ..$ vital   : logi TRUE
#>   ..$ groups  : num 1
#>   ..- attr(*, "class")= chr [1:2] "param.icm" "list"
#>  $ control:List of 11
#>   ..$ type       : chr "SIR"
#>   ..$ nsteps     : num 300
#>   ..$ nsims      : num 10
#>   ..$ rec.rand   : logi TRUE
#>   ..$ a.rand     : logi TRUE
#>   ..$ d.rand     : logi TRUE
#>   ..$ verbose    : logi FALSE
#>   ..$ verbose.int: num 0
#>   ..$ skip.check : logi FALSE
#>   ..$ bi.mods    : chr [1:6] "initialize.FUN" "infection.FUN" "recovery.FUN" "departures.FUN" ...
#>   ..$ user.mods  : chr(0) 
#>   ..- attr(*, "class")= chr [1:2] "control.icm" "list"
#>  $ epi    :List of 10
#>   ..$ s.num  :'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : int [1:300] 900 886 870 849 824 805 790 767 746 717 ...
#>   .. ..$ sim2 : int [1:300] 900 891 874 866 850 827 810 792 764 731 ...
#>   .. ..$ sim3 : int [1:300] 900 882 859 840 817 799 773 749 711 677 ...
#>   .. ..$ sim4 : int [1:300] 900 883 860 834 813 790 775 747 724 690 ...
#>   .. ..$ sim5 : int [1:300] 900 884 870 839 817 790 761 738 708 681 ...
#>   .. ..$ sim6 : int [1:300] 900 889 873 854 830 807 789 767 738 715 ...
#>   .. ..$ sim7 : int [1:300] 900 890 866 852 844 829 797 772 747 725 ...
#>   .. ..$ sim8 : int [1:300] 900 885 869 849 834 816 793 757 733 701 ...
#>   .. ..$ sim9 : int [1:300] 900 884 868 841 815 786 758 737 703 664 ...
#>   .. ..$ sim10: int [1:300] 900 882 866 851 836 823 805 779 748 717 ...
#>   ..$ i.num  :'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : int [1:300] 100 108 121 136 158 176 184 201 220 239 ...
#>   .. ..$ sim2 : int [1:300] 100 106 122 126 141 160 175 192 214 241 ...
#>   .. ..$ sim3 : int [1:300] 100 116 136 154 172 188 209 229 259 286 ...
#>   .. ..$ sim4 : int [1:300] 100 113 131 155 172 193 205 231 250 281 ...
#>   .. ..$ sim5 : int [1:300] 100 113 124 152 172 193 220 239 260 279 ...
#>   .. ..$ sim6 : int [1:300] 100 108 121 137 158 174 190 209 238 257 ...
#>   .. ..$ sim7 : int [1:300] 100 105 124 132 140 150 178 197 212 230 ...
#>   .. ..$ sim8 : int [1:300] 100 114 126 144 156 169 187 220 240 265 ...
#>   .. ..$ sim9 : int [1:300] 100 113 128 147 170 191 214 231 261 298 ...
#>   .. ..$ sim10: int [1:300] 100 113 127 139 154 166 182 204 229 260 ...
#>   ..$ num    :'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim2 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim3 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim4 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim5 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim6 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim7 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim8 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim9 : int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   .. ..$ sim10: int [1:300] 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 ...
#>   ..$ r.num  :'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : int [1:300] 0 6 9 15 18 19 26 32 34 44 ...
#>   .. ..$ sim2 : int [1:300] 0 3 4 8 9 13 15 16 22 28 ...
#>   .. ..$ sim3 : int [1:300] 0 2 5 6 11 13 18 22 30 37 ...
#>   .. ..$ sim4 : int [1:300] 0 4 9 11 15 17 20 22 26 29 ...
#>   .. ..$ sim5 : int [1:300] 0 3 6 9 11 17 19 23 32 40 ...
#>   .. ..$ sim6 : int [1:300] 0 3 6 9 12 19 21 24 24 28 ...
#>   .. ..$ sim7 : int [1:300] 0 5 10 16 16 21 25 31 41 45 ...
#>   .. ..$ sim8 : int [1:300] 0 1 5 7 10 15 20 23 27 34 ...
#>   .. ..$ sim9 : int [1:300] 0 3 4 12 15 23 28 32 36 38 ...
#>   .. ..$ sim10: int [1:300] 0 5 7 10 10 11 13 17 23 23 ...
#>   ..$ si.flow:'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : num [1:300] 0 14 16 21 25 19 15 23 21 29 ...
#>   .. ..$ sim2 : num [1:300] 0 9 17 8 16 23 17 18 28 33 ...
#>   .. ..$ sim3 : num [1:300] 0 18 23 19 23 18 26 24 38 34 ...
#>   .. ..$ sim4 : num [1:300] 0 17 23 26 21 23 15 28 23 34 ...
#>   .. ..$ sim5 : num [1:300] 0 16 14 31 22 27 29 23 30 27 ...
#>   .. ..$ sim6 : num [1:300] 0 11 16 19 24 23 18 22 29 23 ...
#>   .. ..$ sim7 : num [1:300] 0 10 24 14 8 15 32 25 25 22 ...
#>   .. ..$ sim8 : num [1:300] 0 15 16 20 15 18 23 36 24 32 ...
#>   .. ..$ sim9 : num [1:300] 0 16 16 27 26 29 28 21 34 39 ...
#>   .. ..$ sim10: num [1:300] 0 18 16 15 15 13 18 26 31 31 ...
#>   ..$ ir.flow:'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : num [1:300] 0 6 3 6 3 1 7 6 2 10 ...
#>   .. ..$ sim2 : num [1:300] 0 3 1 4 1 4 2 1 6 6 ...
#>   .. ..$ sim3 : num [1:300] 0 2 3 1 5 2 5 4 8 7 ...
#>   .. ..$ sim4 : num [1:300] 0 4 5 2 4 2 3 2 4 3 ...
#>   .. ..$ sim5 : num [1:300] 0 3 3 3 2 6 2 4 9 8 ...
#>   .. ..$ sim6 : num [1:300] 0 3 3 3 3 7 2 3 0 4 ...
#>   .. ..$ sim7 : num [1:300] 0 5 5 6 0 5 4 6 10 4 ...
#>   .. ..$ sim8 : num [1:300] 0 1 4 2 3 5 5 3 4 7 ...
#>   .. ..$ sim9 : num [1:300] 0 3 1 8 3 8 5 4 4 2 ...
#>   .. ..$ sim10: num [1:300] 0 5 2 3 0 1 2 4 6 0 ...
#>   ..$ ds.flow:'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim2 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim3 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim4 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim5 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim6 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim7 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim8 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim9 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim10: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ di.flow:'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim2 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim3 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim4 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim5 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim6 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim7 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim8 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim9 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim10: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ dr.flow:'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim2 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim3 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim4 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim5 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim6 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim7 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim8 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim9 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim10: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ a.flow :'data.frame':	300 obs. of  10 variables:
#>   .. ..$ sim1 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim2 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim3 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim4 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim5 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim6 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim7 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim8 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim9 : num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. ..$ sim10: num [1:300] 0 0 0 0 0 0 0 0 0 0 ...
#>  - attr(*, "class")= chr "icm"

## Look at SIR values
head(EpiModel_icm$epi$s.num)

#>   sim1 sim2 sim3 sim4 sim5 sim6 sim7 sim8 sim9 sim10
#> 1  900  900  900  900  900  900  900  900  900   900
#> 2  886  891  882  883  884  889  890  885  884   882
#> 3  870  874  859  860  870  873  866  869  868   866
#> 4  849  866  840  834  839  854  852  849  841   851
#> 5  824  850  817  813  817  830  844  834  815   836
#> 6  805  827  799  790  790  807  829  816  786   823