agents_sims Example output from SIR simulations

This simulated data is an example of the output from the function simulate_SIR_agents() when output_format = "array". It is a 50 x 3 x 188 array where entry (i,j,k) loks at the ith simulation, the jth statistic, and the kth agent. Each column is asssociated with 1 agent and is a "sufficient" statistic for each agent's infection. Each agent's infection is uniquely identified by an initial state, max time before infection (or time T-1), and max time before recovery (or time T-1). Note time T-1 is the final recorded time in the SIR model (in this case T = 50).

agents_sims

Format

An array with dimension 50 x 3 x 188 data frame with 188 rows and 3 columns. Below are the descriptions of each dimension

sim

Simulation number is the first dimension. They are named between 1 and 50

agent_stat

This is the sufficient statistic for each agent as described above. The names of the statistics are init_state, max_time_S, and max_time_I

agent_id

the ID of the agent

Examples

## show first simulation of first 10 agents
agents_sims[1, , 1:10]

#>             agent_id
#> U_stat       id_1 id_2 id_3 id_4 id_5 id_6 id_7 id_8 id_9 id_10
#>   init_state    0    0    0    0    0    0    0    0    0     0
#>   max_time_S   NA   NA   20   NA   NA   NA   10   NA   NA    15
#>   max_time_I   NA   NA   28   NA   NA   NA   16   NA   NA    19