Calculate filament depth (relative to distance depth)

Calculate filament depth (relative to distance depth)

filament_distance_depth(
  grouped_df,
  data_columns = NULL,
  dist_list_func = "auto",
  depth_func = distance_depth_function
)

Arguments

grouped_df
grouped_df data.frame object (assumed rows per filament are ordered) - grouped per each filament

data_columns
columns of data.frame that relate to the filament's coordinates in euclidean space. The input should look like something like c(S,I,R) or c("S", "I", "R"). If the input is NULL this function will treat this like all non-group columns.

dist_list_func
function to calculate distance between each filament. The default is "auto", which is selects either dist_matrix_innersq_direction if all filaments are the same length -or- preforms a 13 point compression of each filament using filament_compression.

depth_func
function to distances observed between filaments, default is distance_depth_function

Value

depth_vector (of depth related to each filament)