Summary on fragments

This function outputs a summary on fragments using metrics generated by Picard. Provides the number of mapped fragments, duplication rate and number of unique fragments.

fragment_info(picard_list, namelist)

Arguments

picard_list

List of duplication metrics generated by Picard in data frame. Data frames must be listed using `list()` even for one sample (see example below). To import Picard duplication metrics (.txt file) into R, use `picard <- read.table("/path/to/picard/output", header = TRUE, fill = TRUE)`

namelist

List of names in the same order as list of Picard outputs. Add multiple by `c()`

Value

A table summarizing metrics on fragments.

Examples

library(EpiCompare)
data(CnT_H3K27ac_picard) # load example picard output
fragment_info(picard_list = list(CnT_H3K27ac_picard),
              namelist = "CUT&Tag")
#>         Mapped_Fragments Duplication_Rate Unique_Fragments
#> CUT&Tag          2293087            20.18          1830390