Make file names

Support function for gather_files.

gather_files_names(paths, type, nfcore_cutandrun, verbose = TRUE)

Arguments

paths

Character vector of file paths.

type

File type to search for. Options include:

"<pattern>"Finds files matching an arbitrary regex pattern specified by user.

"peaks.stringent"Finds files ending in "*.stringent.bed$"

"peaks.consensus"Finds files ending in "*.consensus.peaks.bed$"

"peaks.consensus.filtered" Finds files ending in"*.consensus.peaks.filtered.awk.bed$"

"picard"Finds files ending in "*.target.markdup.MarkDuplicates.metrics.txt$"

nfcore_cutandrun

Whether the files were generated by the nf-core/cutandrun Nextflow pipeline. If TRUE, can use the standardised folder structure to automatically generate more descriptive file names with sample IDs.

verbose

Print messages.

Value

Named character vector.