Get BiocParallel parameters

Get (and optionally register) BiocParallel parameter (BPPARAM). SnowParam is the default function as it tends to be more robust. However, because it doesn't work on Windows, this function automatically detected the Operating System and switches to SerialParam as needed.

get_bpparam(
  workers,
  progressbar = workers > 1,
  use_snowparam = TRUE,
  register_now = FALSE
)

Arguments

workers

Number of threads to parallelize across.

progressbar

logical(1) Enable progress bar (based on plyr:::progress_text).

use_snowparam

Whether to use SnowParam (default: TRUE) or MulticoreParam (FALSE) when parallelising across multiple workers.

register_now

Register the cores now with register (TRUE), or simply return the BPPARAM object (default: FALSE).

Value

BPPARAM