Measles in Hagelloch, Germany, 1861 (agent format)

This data is the agent formulation of 188 cases of measles among children in the German city of Hagelloch, 1861, and is a summarization of the data in hagelloch_raw. Each column is asssociated with 1 agent and is a "sufficient" statistic for each agent's infection. Each agent's infection is uniquely identified by an initial state, max time before infection (or time T), and max time before recovery (or time T). Note time T is the final recorded time in the SIR model (in this case T = 94).

hagelloch_agents

Format

A data frame with 188 rows and 3 columns. Below are the description of each column.

init_state

Initial state (factor) for individual (at time t = 0). For the states, 0 = S, 1 = I, 2 = R.

max_time_S

maximum time individual was suspectable (S)

max_time_I

maximum time individual was infected (I)

Examples

## show first few cases
head(hagelloch_agents)

#>   init_state max_time_S max_time_I
#> 1          0         22         30
#> 2          0         24         32
#> 3          0         29         37
#> 4          0         28         33
#> 5          0         23         32
#> 6          0         27         34