Measles in Hagelloch, Germany, 1861

The data comprise of 188 cases of measles among children in the German city of Hagelloch, 1861. The data were originally collected by Dr. Albert Pfeilsticker (1863) and augmented and re-analysed by Dr. Heike Oesterle (1992). This data comes from the package outbreaks.

hallegoch_raw contains the exact data as desired in this document. hallegoch_raw2 has a few rows that have no recorded times tR and tI which the user can image means the individual never reached the Recovery or Infected state (note if tI is NA, then so is tR). Additionally, hallegoch_raw2 has 2 individuals that are in recovery at the start of the observational approach, and as such have tR, tI, tPRE, and tERU that are negative.

hagelloch_raw

hagelloch_raw2

hagelloch_aug_births

Format

A data frame with 188 rows and 12 columns

PN

Patient number (Case ID) per individual

NAME

Patient's last name

FN

Family number (individuals with matching FN are in the same family)

HN

Household number ((individuals with matching HN are in the same household)

AGE

Age in years (fractions ignored)

SEX

Gender of the individual (factor: female; male)

PRO

Date of onset of prodromal symptoms

ERU

Date of onset of rash (not 100 sure of this one)

CL

School class (factor: preschool; 1st class; 2nd class )

DEAD

Date of death (with missings)

IFTO

number of patient who is the putative source of infection (0 = unknown)

SI

serial interval = number of days between dates of prodromes of infection source and infected person

C

Complications (factor: "no complications", or what complication)

PR

duration of prodromes in days

CA

number of cases in family

NI

number of initial cases

GE

generation number of the case

TD

day of max. fever (days after rush)

TM

max.fever (degree Celsius)

x.loc

x coordinate of house (in metres). Scaling in metres is obtained by multiplying the original coordinates by 2.5 (see details in Neal and Roberts (2004))

y.loc

y coordinate of house (in metres). See x.loc above

tPRO

exact time of onset of prodromal symptoms (relative to start of outbreak)

tERU

exact time of onset of rash (relative to start of outbreak)

tDEAD

time of death (relative to start of outbreak) (NA implies recovered)

tR

imputed time when individual entered into recovery

tI

imputed time when individual was infected

tBORN

Imputed time when individual was born. (Only for hagelloch_aug_births

An object of class agents_df (inherits from agents_df, data.frame) with 188 rows and 26 columns.

An object of class agents_df (inherits from data.frame) with 193 rows and 27 columns.

Source

https://rdrr.io/rforge/surveillance/man/hagelloch.html

Examples

## show first few cases
head(hagelloch_raw); head(hagelloch_raw2)

#>   PN    NAME FN HN AGE    SEX        PRO        ERU        CL DEAD IFTO SI
#> 1  1 Mueller 41 61   7 female 1861-11-21 1861-11-25 1st class <NA>   45 10
#> 2  2 Mueller 41 61   6 female 1861-11-23 1861-11-27 1st class <NA>   45 12
#> 3  3 Mueller 41 61   4 female 1861-11-28 1861-12-02 preschool <NA>  172  9
#> 4  4 Seibold 61 62  13   male 1861-11-27 1861-11-28 2nd class <NA>  180 10
#> 5  5  Motzer 42 63   8 female 1861-11-22 1861-11-27 1st class <NA>   45 11
#> 6  6  Motzer 42 63  12   male 1861-11-26 1861-11-29 2nd class <NA>  180  9
#>                  C PR CA NI GE TD   TM x.loc y.loc     tPRO     tERU tDEAD
#> 1  no complicatons  4  4  3  1 NA   NA 142.5 100.0 22.71242 26.22541    NA
#> 2  no complicatons  4  4  3  1  3 40.3 142.5 100.0 24.21169 28.79112    NA
#> 3  no complicatons  4  4  3  2  1 40.5 142.5 100.0 29.59102 33.69121    NA
#> 4  no complicatons  1  1  1  1  3 40.7 165.0 102.5 28.11698 29.02866    NA
#> 5  no complicatons  5  3  2  1 NA   NA 145.0 120.0 23.05953 28.41510    NA
#> 6 bronchopneumonia  3  3  2  1  2 40.7 145.0 120.0 27.95444 30.23918    NA
#>         tR       tI
#> 1 29.22541 21.71242
#> 2 31.79112 23.21169
#> 3 36.69121 28.59102
#> 4 32.02866 27.11698
#> 5 31.41510 22.05953
#> 6 33.23918 26.95444

#>   PN    NAME FN HN AGE    SEX        PRO        ERU        CL DEAD IFTO SI
#> 1  1 Mueller 41 61   7 female 1861-11-21 1861-11-25 1st class <NA>   45 10
#> 2  2 Mueller 41 61   6 female 1861-11-23 1861-11-27 1st class <NA>   45 12
#> 3  3 Mueller 41 61   4 female 1861-11-28 1861-12-02 preschool <NA>  172  9
#> 4  4 Seibold 61 62  13   male 1861-11-27 1861-11-28 2nd class <NA>  180 10
#> 5  5  Motzer 42 63   8 female 1861-11-22 1861-11-27 1st class <NA>   45 11
#> 6  6  Motzer 42 63  12   male 1861-11-26 1861-11-29 2nd class <NA>  180  9
#>                  C PR CA NI GE TD   TM x.loc y.loc     tPRO     tERU tDEAD
#> 1  no complicatons  4  4  3  1 NA   NA 142.5 100.0 22.71242 26.22541    NA
#> 2  no complicatons  4  4  3  1  3 40.3 142.5 100.0 24.21169 28.79112    NA
#> 3  no complicatons  4  4  3  2  1 40.5 142.5 100.0 29.59102 33.69121    NA
#> 4  no complicatons  1  1  1  1  3 40.7 165.0 102.5 28.11698 29.02866    NA
#> 5  no complicatons  5  3  2  1 NA   NA 145.0 120.0 23.05953 28.41510    NA
#> 6 bronchopneumonia  3  3  2  1  2 40.7 145.0 120.0 27.95444 30.23918    NA
#>         tR       tI
#> 1 29.22541 21.71242
#> 2 31.79112 23.21169
#> 3 36.69121 28.59102
#> 4 32.02866 27.11698
#> 5 31.41510 22.05953
#> 6 33.23918 26.95444