Import narrow peaks

Import narrowPeak files generated by MACS.

import_narrowPeak(
  path,
  as_file = file.path("https://www.encodeproject.org/documents",
    "948203bb-8eb2-42a2-8b12-1c10f356c998", "@@download/attachment", "narrowPeak.as"),
  extraCols = c(signalValue = "numeric", pValue = "numeric", qValue = "numeric", peak =
    "integer")
)

Source

Solution from Bioconductor forum

Arguments

path

Path to local narrowPeak file or remote URL.

as_file

Metadata describing each of the column names in narrowPeak files. Provided by ENCODE.

extraCols

What vector type to assign each of the extra columns to.

Value

GRanges object.

Details

Example dataset from: https://www.encodeproject.org/files/ENCFF044JNJ/. See encode_H3K27ac for more details.

Examples

# path <- file.path("https://www.encodeproject.org/files",
#                   "ENCFF044JNJ/@download/ENCFF044JNJ.bed.gz")

path <- EpiCompare::write_example_peaks(datasets="encode_H3K27ac")
#> Writing ==> /tmp/RtmpVafuVX/processed_results/encode_H3K27ac.narrowPeaks.bed
peaks <- EpiCompare::import_narrowPeak(path=path)
#> 5,142 peaks imported.