Run delta ball analysis

Run delta-ball analysis to obtain an outline points for delta-ball covering where points are in 2d space. We first find the minimum delta such that all balls centered at each point in the data set is touching at least 1 other ball (see get_delta for more information). This function then creates a geometric objects that trying to represent the covering of all these delta balls. Specifically we use geometric properties to find the points on the "outside" of this covering and return a set of lines that create a "boundary" of shorts. Intuition from "Computing Polygonal Surfaces from Unions of Balls" by Tam and Heidrich was used in this function.

inner_delta_ball_wrapper(
  data_raw,
  xy_columns = c("x", "y"),
  n_steps = 100,
  remove_duplicates = F
)

Arguments

data_raw
data frame with center points of balls

xy_columns
columns of data.frame that relate to the points's coordinates in euclidean space. The input should look like something like c(x,y) or c("x","y").

n_steps
number of equidistance points along the line, past delta on both sides, that will be checked to approximate all points along the line

remove_duplicates
boolean if need to remove duplicates in data_raw

Value

data frame of exterior lines (not ordered)

Details

This function (renamed as delta_ball_wrapper) is shared with TCpredictionbands on github: TCpredictionbands.