Generate heatmap of percentage overlap

This function generates a heatmap showing percentage of overlapping peaks between peak files.

overlap_heatmap(peaklist, interact = TRUE)

Arguments

peaklist

A list of peak files as GRanges object. Files must be listed using `list()` and named using `names()` If not named, default file names will be assigned.

interact

Default TRUE. By default heatmap is interactive. If FALSE, heatmap is static.

Value

An interactive heatmap

Examples

library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object

peaks <- list(encode_H3K27ac, CnT_H3K27ac) # create list
names(peaks) <- c("encode", "CnT") # set names

overlap_heatmap(peaklist = peaks)