Create ChIPseeker annotation plot

This function annotates peaks using `annotatePeak` from `ChIPseeker` package. It outputs functional annotation of each peak file in a barplot.

plot_ChIPseeker_annotation(peaklist, annotation)

Arguments

peaklist

A list of peak files as GRanges object. Files must be listed using `list()` and named using `names()` If not named, default file names will be assigned.

annotation

A TxDb annotation object from Bioconductor.

Value

barplot

Examples

data("CnT_H3K27ac") # example dataset as GRanges object
data("CnR_H3K27ac") # example dataset as GRanges object

peaks <- list(CnT_H3K27ac, CnR_H3K27ac) # create a list
names(peaks) <- c("CnT", "CnR") # set names

## not run
# txdb<-TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
# my_plot <- plot_ChIPseeker_annotation(peaklist = peaks
#                                       annotation = txdb)