Generate enrichment analysis plots

This function runs KEGG and GO enrichment analysis of peak files and generates dot plots.

plot_enrichment(
  peaklist,
  txdb = NULL,
  tss_distance = c(-3000, 3000),
  pvalueCutoff = 0.05,
  interact = FALSE,
  verbose = TRUE
)

Arguments

peaklist

A list of peak files as GRanges object. Files must be listed and named using list(). e.g. list("name1"=file1, "name2"=file2). If not named, default file names will be assigned.

txdb

A TxDb annotation object from Bioconductor.

tss_distance

A vector specifying the distance upstream and downstream around transcription start sites (TSS). The default value is c(-3000,3000); meaning peak frequency 3000bp upstream and downstream of TSS will be displayed.

pvalueCutoff

P-value cutoff, passed to compareCluster.

interact

Default TRUE. By default, plots are interactive. If set FALSE, all plots in the report will be static.

verbose

Print messages.

Value

KEGG and GO dot plots

Examples

### Load Data ###
data("CnT_H3K27ac") # example peakfile GRanges object
data("CnR_H3K27ac") # example peakfile GRanges object 
### Create Named Peaklist ###
peaklist <- list("CnT"=CnT_H3K27ac, "CnR"=CnR_H3K27ac) 
enrich_res <- plot_enrichment(peaklist = peaklist, pvalueCutoff=1,
                              tss_distance = c(-50,50)) 
#> --- Running plot_enrichment() ---
#> 
#> Warning: >> TxDb is not specified, use 'TxDb.Hsapiens.UCSC.hg19.knownGene' by default...
#> Warning: >> TxDb is not specified, use 'TxDb.Hsapiens.UCSC.hg19.knownGene' by default...
#> + Running clusterProfiler::compareCluster for KEGG.
#> + Running clusterProfiler::compareCluster for GO.
#> plot_enrichment(): Done in 23.6s.