Example SIR simulation output of class ' pomp' from the pomp package

The data is an example of simulation output from functions from the pomp package by King et al. The code for how this data was generated can be found in the data-raw folder. We also make available the other possible output formats from pomp. See pomp_df and pomp_arr

pomp_pomp

Format

The object is a 'pompList' which contains essentially all information from a pomp simulation.

Examples

data("pomp_pomp")
str(head(pomp_pomp))

#> Formal class 'pompList' [package "pomp"] with 1 slot
#>   ..@ .Data:List of 6
#>   .. ..$ :Formal class 'pomp' [package "pomp"] with 19 slots
#>   .. .. .. ..@ data     : num [1, 1:101] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr "cases"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ times    : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. .. .. ..@ t0       : num 0
#>   .. .. .. ..@ timename : chr "time"
#>   .. .. .. ..@ rinit    :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rinit"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rinit"
#>   .. .. .. ..@ rprocess :Formal class 'eulerRprocPlugin' [package "pomp"] with 6 slots
#>   .. .. .. .. .. ..@ delta.t : num 1
#>   .. .. .. .. .. ..@ csnippet: logi TRUE
#>   .. .. .. .. .. ..@ slotname: chr "step.fun"
#>   .. .. .. .. .. ..@ type    : int 3
#>   .. .. .. .. .. ..@ step.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr "__pomp_stepfn"
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "step.fun"
#>   .. .. .. .. .. ..@ rate.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "rate.fun"
#>   .. .. .. ..@ dprocess :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprocess"
#>   .. .. .. ..@ rmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int 0
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rmeasure"
#>   .. .. .. ..@ dmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_dmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dmeasure"
#>   .. .. .. ..@ rprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rprior"
#>   .. .. .. ..@ dprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprior"
#>   .. .. .. ..@ skeleton :Formal class 'skelPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ csnippet: logi FALSE
#>   .. .. .. .. .. ..@ type    : int 0
#>   .. .. .. .. .. ..@ skel.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "skeleton"
#>   .. .. .. ..@ partrans :Formal class 'partransPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ has : logi FALSE
#>   .. .. .. .. .. ..@ to  : NULL
#>   .. .. .. .. .. ..@ from: NULL
#>   .. .. .. ..@ params   : Named num [1:9] 1000 0.1 0.03 0 0 0 950 50 0
#>   .. .. .. .. ..- attr(*, "names")= chr [1:9] "popsize" "Beta" "gamma" "mu" ...
#>   .. .. .. ..@ states   : num [1:4, 1:101] 950 50 0 0 946 52 2 4 944 54 ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ covar    :Formal class 'covartable' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ times: num(0) 
#>   .. .. .. .. .. ..@ table: num[0 , 0 ] 
#>   .. .. .. .. .. ..@ order: int 1
#>   .. .. .. ..@ accumvars: chr "H"
#>   .. .. .. ..@ solibs   :List of 1
#>   .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. ..$ name: chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..$ dir : NULL
#>   .. .. .. .. .. ..$ src : chr "/* pomp C snippet file: pomp_39d2e6d6d1c74c96cbaa536ecd6833bb */\n/* Time: 2019-10-29 11:37:03.381 -0400 */\n/*"| __truncated__
#>   .. .. .. ..@ userdata : list()
#>   .. ..$ :Formal class 'pomp' [package "pomp"] with 19 slots
#>   .. .. .. ..@ data     : num [1, 1:101] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr "cases"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ times    : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. .. .. ..@ t0       : num 0
#>   .. .. .. ..@ timename : chr "time"
#>   .. .. .. ..@ rinit    :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rinit"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rinit"
#>   .. .. .. ..@ rprocess :Formal class 'eulerRprocPlugin' [package "pomp"] with 6 slots
#>   .. .. .. .. .. ..@ delta.t : num 1
#>   .. .. .. .. .. ..@ csnippet: logi TRUE
#>   .. .. .. .. .. ..@ slotname: chr "step.fun"
#>   .. .. .. .. .. ..@ type    : int 3
#>   .. .. .. .. .. ..@ step.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr "__pomp_stepfn"
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "step.fun"
#>   .. .. .. .. .. ..@ rate.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "rate.fun"
#>   .. .. .. ..@ dprocess :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprocess"
#>   .. .. .. ..@ rmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int 0
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rmeasure"
#>   .. .. .. ..@ dmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_dmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dmeasure"
#>   .. .. .. ..@ rprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rprior"
#>   .. .. .. ..@ dprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprior"
#>   .. .. .. ..@ skeleton :Formal class 'skelPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ csnippet: logi FALSE
#>   .. .. .. .. .. ..@ type    : int 0
#>   .. .. .. .. .. ..@ skel.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "skeleton"
#>   .. .. .. ..@ partrans :Formal class 'partransPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ has : logi FALSE
#>   .. .. .. .. .. ..@ to  : NULL
#>   .. .. .. .. .. ..@ from: NULL
#>   .. .. .. ..@ params   : Named num [1:9] 1000 0.1 0.03 0 0 0 950 50 0
#>   .. .. .. .. ..- attr(*, "names")= chr [1:9] "popsize" "Beta" "gamma" "mu" ...
#>   .. .. .. ..@ states   : num [1:4, 1:101] 950 50 0 0 947 51 2 3 944 51 ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ covar    :Formal class 'covartable' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ times: num(0) 
#>   .. .. .. .. .. ..@ table: num[0 , 0 ] 
#>   .. .. .. .. .. ..@ order: int 1
#>   .. .. .. ..@ accumvars: chr "H"
#>   .. .. .. ..@ solibs   :List of 1
#>   .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. ..$ name: chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..$ dir : NULL
#>   .. .. .. .. .. ..$ src : chr "/* pomp C snippet file: pomp_39d2e6d6d1c74c96cbaa536ecd6833bb */\n/* Time: 2019-10-29 11:37:03.381 -0400 */\n/*"| __truncated__
#>   .. .. .. ..@ userdata : list()
#>   .. ..$ :Formal class 'pomp' [package "pomp"] with 19 slots
#>   .. .. .. ..@ data     : num [1, 1:101] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr "cases"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ times    : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. .. .. ..@ t0       : num 0
#>   .. .. .. ..@ timename : chr "time"
#>   .. .. .. ..@ rinit    :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rinit"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rinit"
#>   .. .. .. ..@ rprocess :Formal class 'eulerRprocPlugin' [package "pomp"] with 6 slots
#>   .. .. .. .. .. ..@ delta.t : num 1
#>   .. .. .. .. .. ..@ csnippet: logi TRUE
#>   .. .. .. .. .. ..@ slotname: chr "step.fun"
#>   .. .. .. .. .. ..@ type    : int 3
#>   .. .. .. .. .. ..@ step.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr "__pomp_stepfn"
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "step.fun"
#>   .. .. .. .. .. ..@ rate.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "rate.fun"
#>   .. .. .. ..@ dprocess :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprocess"
#>   .. .. .. ..@ rmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int 0
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rmeasure"
#>   .. .. .. ..@ dmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_dmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dmeasure"
#>   .. .. .. ..@ rprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rprior"
#>   .. .. .. ..@ dprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprior"
#>   .. .. .. ..@ skeleton :Formal class 'skelPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ csnippet: logi FALSE
#>   .. .. .. .. .. ..@ type    : int 0
#>   .. .. .. .. .. ..@ skel.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "skeleton"
#>   .. .. .. ..@ partrans :Formal class 'partransPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ has : logi FALSE
#>   .. .. .. .. .. ..@ to  : NULL
#>   .. .. .. .. .. ..@ from: NULL
#>   .. .. .. ..@ params   : Named num [1:9] 1000 0.1 0.03 0 0 0 950 50 0
#>   .. .. .. .. ..- attr(*, "names")= chr [1:9] "popsize" "Beta" "gamma" "mu" ...
#>   .. .. .. ..@ states   : num [1:4, 1:101] 950 50 0 0 943 56 1 7 939 60 ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ covar    :Formal class 'covartable' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ times: num(0) 
#>   .. .. .. .. .. ..@ table: num[0 , 0 ] 
#>   .. .. .. .. .. ..@ order: int 1
#>   .. .. .. ..@ accumvars: chr "H"
#>   .. .. .. ..@ solibs   :List of 1
#>   .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. ..$ name: chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..$ dir : NULL
#>   .. .. .. .. .. ..$ src : chr "/* pomp C snippet file: pomp_39d2e6d6d1c74c96cbaa536ecd6833bb */\n/* Time: 2019-10-29 11:37:03.381 -0400 */\n/*"| __truncated__
#>   .. .. .. ..@ userdata : list()
#>   .. ..$ :Formal class 'pomp' [package "pomp"] with 19 slots
#>   .. .. .. ..@ data     : num [1, 1:101] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr "cases"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ times    : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. .. .. ..@ t0       : num 0
#>   .. .. .. ..@ timename : chr "time"
#>   .. .. .. ..@ rinit    :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rinit"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rinit"
#>   .. .. .. ..@ rprocess :Formal class 'eulerRprocPlugin' [package "pomp"] with 6 slots
#>   .. .. .. .. .. ..@ delta.t : num 1
#>   .. .. .. .. .. ..@ csnippet: logi TRUE
#>   .. .. .. .. .. ..@ slotname: chr "step.fun"
#>   .. .. .. .. .. ..@ type    : int 3
#>   .. .. .. .. .. ..@ step.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr "__pomp_stepfn"
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "step.fun"
#>   .. .. .. .. .. ..@ rate.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "rate.fun"
#>   .. .. .. ..@ dprocess :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprocess"
#>   .. .. .. ..@ rmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int 0
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rmeasure"
#>   .. .. .. ..@ dmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_dmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dmeasure"
#>   .. .. .. ..@ rprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rprior"
#>   .. .. .. ..@ dprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprior"
#>   .. .. .. ..@ skeleton :Formal class 'skelPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ csnippet: logi FALSE
#>   .. .. .. .. .. ..@ type    : int 0
#>   .. .. .. .. .. ..@ skel.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "skeleton"
#>   .. .. .. ..@ partrans :Formal class 'partransPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ has : logi FALSE
#>   .. .. .. .. .. ..@ to  : NULL
#>   .. .. .. .. .. ..@ from: NULL
#>   .. .. .. ..@ params   : Named num [1:9] 1000 0.1 0.03 0 0 0 950 50 0
#>   .. .. .. .. ..- attr(*, "names")= chr [1:9] "popsize" "Beta" "gamma" "mu" ...
#>   .. .. .. ..@ states   : num [1:4, 1:101] 950 50 0 0 946 53 1 4 942 57 ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ covar    :Formal class 'covartable' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ times: num(0) 
#>   .. .. .. .. .. ..@ table: num[0 , 0 ] 
#>   .. .. .. .. .. ..@ order: int 1
#>   .. .. .. ..@ accumvars: chr "H"
#>   .. .. .. ..@ solibs   :List of 1
#>   .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. ..$ name: chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..$ dir : NULL
#>   .. .. .. .. .. ..$ src : chr "/* pomp C snippet file: pomp_39d2e6d6d1c74c96cbaa536ecd6833bb */\n/* Time: 2019-10-29 11:37:03.381 -0400 */\n/*"| __truncated__
#>   .. .. .. ..@ userdata : list()
#>   .. ..$ :Formal class 'pomp' [package "pomp"] with 19 slots
#>   .. .. .. ..@ data     : num [1, 1:101] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr "cases"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ times    : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. .. .. ..@ t0       : num 0
#>   .. .. .. ..@ timename : chr "time"
#>   .. .. .. ..@ rinit    :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rinit"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rinit"
#>   .. .. .. ..@ rprocess :Formal class 'eulerRprocPlugin' [package "pomp"] with 6 slots
#>   .. .. .. .. .. ..@ delta.t : num 1
#>   .. .. .. .. .. ..@ csnippet: logi TRUE
#>   .. .. .. .. .. ..@ slotname: chr "step.fun"
#>   .. .. .. .. .. ..@ type    : int 3
#>   .. .. .. .. .. ..@ step.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr "__pomp_stepfn"
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "step.fun"
#>   .. .. .. .. .. ..@ rate.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "rate.fun"
#>   .. .. .. ..@ dprocess :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprocess"
#>   .. .. .. ..@ rmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int 0
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rmeasure"
#>   .. .. .. ..@ dmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_dmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dmeasure"
#>   .. .. .. ..@ rprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rprior"
#>   .. .. .. ..@ dprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprior"
#>   .. .. .. ..@ skeleton :Formal class 'skelPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ csnippet: logi FALSE
#>   .. .. .. .. .. ..@ type    : int 0
#>   .. .. .. .. .. ..@ skel.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "skeleton"
#>   .. .. .. ..@ partrans :Formal class 'partransPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ has : logi FALSE
#>   .. .. .. .. .. ..@ to  : NULL
#>   .. .. .. .. .. ..@ from: NULL
#>   .. .. .. ..@ params   : Named num [1:9] 1000 0.1 0.03 0 0 0 950 50 0
#>   .. .. .. .. ..- attr(*, "names")= chr [1:9] "popsize" "Beta" "gamma" "mu" ...
#>   .. .. .. ..@ states   : num [1:4, 1:101] 950 50 0 0 948 52 0 2 943 56 ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ covar    :Formal class 'covartable' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ times: num(0) 
#>   .. .. .. .. .. ..@ table: num[0 , 0 ] 
#>   .. .. .. .. .. ..@ order: int 1
#>   .. .. .. ..@ accumvars: chr "H"
#>   .. .. .. ..@ solibs   :List of 1
#>   .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. ..$ name: chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..$ dir : NULL
#>   .. .. .. .. .. ..$ src : chr "/* pomp C snippet file: pomp_39d2e6d6d1c74c96cbaa536ecd6833bb */\n/* Time: 2019-10-29 11:37:03.381 -0400 */\n/*"| __truncated__
#>   .. .. .. ..@ userdata : list()
#>   .. ..$ :Formal class 'pomp' [package "pomp"] with 19 slots
#>   .. .. .. ..@ data     : num [1, 1:101] NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr "cases"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ times    : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. .. .. ..@ t0       : num 0
#>   .. .. .. ..@ timename : chr "time"
#>   .. .. .. ..@ rinit    :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rinit"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rinit"
#>   .. .. .. ..@ rprocess :Formal class 'eulerRprocPlugin' [package "pomp"] with 6 slots
#>   .. .. .. .. .. ..@ delta.t : num 1
#>   .. .. .. .. .. ..@ csnippet: logi TRUE
#>   .. .. .. .. .. ..@ slotname: chr "step.fun"
#>   .. .. .. .. .. ..@ type    : int 3
#>   .. .. .. .. .. ..@ step.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr "__pomp_stepfn"
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "step.fun"
#>   .. .. .. .. .. ..@ rate.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "rate.fun"
#>   .. .. .. ..@ dprocess :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprocess"
#>   .. .. .. ..@ rmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_rmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int [1:4] 0 1 2 3
#>   .. .. .. .. .. ..@ paramindex: int [1:9] 2 3 5 1 0 4 6 7 8
#>   .. .. .. .. .. ..@ obsindex  : int 0
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rmeasure"
#>   .. .. .. ..@ dmeasure :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr "__pomp_dmeasure"
#>   .. .. .. .. .. ..@ PACKAGE   : chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..@ mode      : int 3
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dmeasure"
#>   .. .. .. ..@ rprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "rprior"
#>   .. .. .. ..@ dprior   :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. ..@ purpose   : chr "dprior"
#>   .. .. .. ..@ skeleton :Formal class 'skelPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ csnippet: logi FALSE
#>   .. .. .. .. .. ..@ type    : int 0
#>   .. .. .. .. .. ..@ skel.fn :Formal class 'pomp_fun' [package "pomp"] with 14 slots
#>   .. .. .. .. .. .. .. ..@ R.fun     :function (...)  
#>   .. .. .. .. .. .. .. ..@ native.fun: chr(0) 
#>   .. .. .. .. .. .. .. ..@ PACKAGE   : chr(0) 
#>   .. .. .. .. .. .. .. ..@ mode      : int 0
#>   .. .. .. .. .. .. .. ..@ address   :<externalptr> 
#>   .. .. .. .. .. .. .. ..@ statenames: chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. .. .. ..@ paramnames: chr [1:9] "gamma" "mu" "theta" "Beta" ...
#>   .. .. .. .. .. .. .. ..@ obsnames  : chr "cases"
#>   .. .. .. .. .. .. .. ..@ covarnames: chr(0) 
#>   .. .. .. .. .. .. .. ..@ stateindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ paramindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ obsindex  : int(0) 
#>   .. .. .. .. .. .. .. ..@ covarindex: int(0) 
#>   .. .. .. .. .. .. .. ..@ purpose   : chr "skeleton"
#>   .. .. .. ..@ partrans :Formal class 'partransPlugin' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ has : logi FALSE
#>   .. .. .. .. .. ..@ to  : NULL
#>   .. .. .. .. .. ..@ from: NULL
#>   .. .. .. ..@ params   : Named num [1:9] 1000 0.1 0.03 0 0 0 950 50 0
#>   .. .. .. .. ..- attr(*, "names")= chr [1:9] "popsize" "Beta" "gamma" "mu" ...
#>   .. .. .. ..@ states   : num [1:4, 1:101] 950 50 0 0 948 51 1 2 942 54 ...
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:4] "S" "I" "R" "H"
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. ..@ covar    :Formal class 'covartable' [package "pomp"] with 3 slots
#>   .. .. .. .. .. ..@ times: num(0) 
#>   .. .. .. .. .. ..@ table: num[0 , 0 ] 
#>   .. .. .. .. .. ..@ order: int 1
#>   .. .. .. ..@ accumvars: chr "H"
#>   .. .. .. ..@ solibs   :List of 1
#>   .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. ..$ name: chr "pomp_39d2e6d6d1c74c96cbaa536ecd6833bb"
#>   .. .. .. .. .. ..$ dir : NULL
#>   .. .. .. .. .. ..$ src : chr "/* pomp C snippet file: pomp_39d2e6d6d1c74c96cbaa536ecd6833bb */\n/* Time: 2019-10-29 11:37:03.381 -0400 */\n/*"| __truncated__
#>   .. .. .. ..@ userdata : list()