Prepare genome builds

Parse the genome_build argument into peaklist_build and reference_build.

prepare_genome_builds(genome_build)

Arguments

genome_build

A named list indicating the human genome build used to generate each of the following inputs:

"peakfiles" : Genome build for the peakfiles input. Assumes genome build is the same for each element in the peakfiles list.

"reference" : Genome build for the reference input.

"blacklist" : Genome build for the blacklist input.

Example input list:
genome_build = list(peakfiles="hg38", reference="hg19", blacklist="hg19")

Alternatively, you can supply a single character string instead of a list. This should only be done in situations where all three inputs (peakfiles, reference, blacklist) are of the same genome build. For example:
genome_build = "hg19"

Value

Named list.