Tidy peakfiles in GRanges

This function filters peak files by removing peaks in blacklisted regions and in non-standard chromosomes. It also checks that the input list of peakfiles is named. If no names are provided, default file names will be used.

tidy_peakfile(peaklist, blacklist)

Arguments

peaklist

A named list of peak files as GRanges object. Objects listed using `list()` and named using `names()`. If not named, default file names are assigned.

blacklist

Peakfile specifying blacklisted regions as GRanges object.

Value

list of GRanges object

Examples

library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object
data("hg19_blacklist") # blacklist region for hg19 genome

peaklist <- list(encode_H3K27ac, CnT_H3K27ac) # list two peakfiles
names(peaklist) <- c("encode", "CnT") # set names

tidy_peakfile(peaklist = peaklist,
              blacklist = hg19_blacklist)
#> $encode
#> GRanges object with 5076 ranges and 7 metadata columns:
#>          seqnames              ranges strand |        name     score
#>             <Rle>           <IRanges>  <Rle> | <character> <integer>
#>      [1]     chr1   10002541-10004163      * |   Peak_2650       170
#>      [2]     chr1 100064339-100065136      * |  Peak_11419        58
#>      [3]     chr1 100065447-100065946      * |  Peak_14209        46
#>      [4]     chr1 100066041-100066635      * |  Peak_33569        19
#>      [5]     chr1   10010590-10010946      * |  Peak_32232        19
#>      ...      ...                 ...    ... .         ...       ...
#>   [5072]     chr1     9970719-9971791      * |   Peak_2896       162
#>   [5073]     chr1     9972220-9972839      * |  Peak_32178        19
#>   [5074]     chr1     9988537-9988798      * |  Peak_52060        14
#>   [5075]     chr1     9988953-9990753      * |   Peak_5018       117
#>   [5076]     chr1      999342-1000149      * |   Peak_7706        84
#>               strand signalValue    pValue    qValue      peak
#>          <character>   <numeric> <numeric> <numeric> <integer>
#>      [1]           .    31.36983  148.0261 144.32048      1348
#>      [2]           .    16.17942   46.3082  43.50269       329
#>      [3]           .    10.24276   35.4649  32.79924       335
#>      [4]           .     5.48637   10.2960   8.09466       456
#>      [5]           .     5.97935   10.7905   8.57204       152
#>      ...         ...         ...       ...       ...       ...
#>   [5072]           .    29.44711 140.57442 136.92255       337
#>   [5073]           .     4.02635  10.83620   8.61721       328
#>   [5074]           .     4.36208   5.95315   3.90332       101
#>   [5075]           .    21.34607  99.30075  95.97326       783
#>   [5076]           .    21.80705  69.82115  66.76275       189
#>   -------
#>   seqinfo: 24 sequences from an unspecified genome; no seqlengths
#> 
#> $CnT
#> GRanges object with 1580 ranges and 7 metadata columns:
#>          seqnames              ranges strand |                   name     score
#>             <Rle>           <IRanges>  <Rle> |            <character> <integer>
#>      [1]     chr1       755479-756187      * | H3K27ac_SRR8383507_q..        76
#>      [2]     chr1       757585-758501      * | H3K27ac_SRR8383507_q..        92
#>      [3]     chr1       762303-763398      * | H3K27ac_SRR8383507_q..       124
#>      [4]     chr1       935483-936365      * | H3K27ac_SRR8383507_q..       141
#>      [5]     chr1       954917-955499      * | H3K27ac_SRR8383507_q..       349
#>      ...      ...                 ...    ... .                    ...       ...
#>   [1576]     chr1 249143558-249144021      * | H3K27ac_SRR8383507_q..       108
#>   [1577]     chr1 249151569-249154018      * | H3K27ac_SRR8383507_q..       431
#>   [1578]     chr1 249166030-249168876      * | H3K27ac_SRR8383507_q..       692
#>   [1579]     chr1 249183459-249185760      * | H3K27ac_SRR8383507_q..       494
#>   [1580]     chr1 249198861-249200960      * | H3K27ac_SRR8383507_q..       494
#>               strand signalValue    pValue    qValue      peak
#>          <character>   <numeric> <numeric> <numeric> <integer>
#>      [1]           .     6.66100   10.2571   7.65796        31
#>      [2]           .     7.49362   11.9091   9.21748       882
#>      [3]           .     9.15887   15.3455  12.48826       822
#>      [4]           .     9.99149   17.1220  14.18825       366
#>      [5]           .    19.15036   38.5216  34.90429       195
#>      ...         ...         ...       ...       ...       ...
#>   [1576]           .     8.32625   13.6066   10.8299       287
#>   [1577]           .    22.48086   46.9336   43.1150      1544
#>   [1578]           .    32.47235   73.5682   69.2412      1491
#>   [1579]           .    24.97873   53.4109   49.4499       435
#>   [1580]           .    24.97873   53.4109   49.4499      1633
#>   -------
#>   seqinfo: 24 sequences from an unspecified genome; no seqlengths
#>