Pull the points from the top (1-alpha)% percent of curves (LIST)

pulling out data points from curves that are ranked highest based on quantile function.

# S3 method for list
top_curves_to_points(
  x,
  alpha,
  tidy_dm,
  quantile_func = distance_depth_function,
  x_names_df,
  ...
)

Arguments

x
list containing curves, with index ordering associated with the x_names_df and rownames of tidy_dm

alpha
the proportion of curves to be removed before presenting all the points together. Takes value in [0, 1.0].

tidy_dm
a tidy_dist_mat distance matrix

quantile_func
function to calculate quantiles via the distance_matrix, we now expect this function to handle tidy_dist_mat objects and have a parameter called df_out which we can set as true. See distance_depth_function.tidy_dist_mat for an example.

x_names_df
group structure associated with the ordering of the items in the list x. Assume the naming structure for tidy_dm is the same as this data frame (can have different ordering).

...
additional parameters to be passed to the quantile_func. Please also see details for more information.

Value

data frame from curves of the top values associated with the quantile_func