Generate heatmap of percentage overlap

This function generates a heatmap showing percentage of overlapping peaks between peak files.

overlap_heatmap(peaklist, namelist, save_dir = NULL)

Arguments

peaklist

A list of peaks as GRanges object. Objects in list using `list()`

namelist

A list of names. Names in list using `c()`

save_dir

Name of file to save the interactive heatmap output as HTML

Value

An interactive heatmap

Examples

library(EpiCompare)
data("encode_H3K27ac") # example dataset as GRanges object
data("CnT_H3K27ac") # example dataset as GRanges object

overlap_heatmap(peaklist = list(encode_H3K27ac, CnT_H3K27ac),
                namelist = c("ENCODE", "CnT"))